lark-12: Refactor naming of new plugin to MetaOption
