+++ lark.services.yml
@@ -32,8 +32,8 @@ services:
     class: Drupal\lark\Service\ExportableFactory
     autowire: true
 
-  Drupal\lark\Service\EntityExportFormPluginManager:
-    class: Drupal\lark\Service\EntityExportFormPluginManager
+  Drupal\lark\Service\MetaOptionManager:
+    class: Drupal\lark\Service\MetaOptionManager
     parent: default_plugin_manager
 
   Drupal\lark\Service\FieldTypeHandlerManagerInterface:
