+++ src/Service/MetaOptionManager.php
@@ -7,19 +7,19 @@ namespace Drupal\lark\Service;
 use Drupal\Core\Cache\CacheBackendInterface;
 use Drupal\Core\Extension\ModuleHandlerInterface;
 use Drupal\Core\Plugin\DefaultPluginManager;
-use Drupal\lark\Attribute\LarkEntityExportForm;
-use Drupal\lark\Plugin\Lark\EntityExportFormInterface;
+use Drupal\lark\Attribute\LarkMetaOption;
+use Drupal\lark\Plugin\Lark\MetaOptionInterface;
 
 /**
- * LarkEntityExportForm plugin manager.
+ * LarkMetaOption plugin manager.
  */
-final class EntityExportFormPluginManager extends DefaultPluginManager {
+final class MetaOptionManager extends DefaultPluginManager {
 
   /**
    * Constructs the object.
    */
   public function __construct(\Traversable $namespaces, CacheBackendInterface $cache_backend, ModuleHandlerInterface $module_handler) {
-    parent::__construct('Plugin/Lark/EntityExportForm', $namespaces, $module_handler, EntityExportFormInterface::class, LarkEntityExportForm::class);
+    parent::__construct('Plugin/Lark/MetaOption', $namespaces, $module_handler, MetaOptionInterface::class, LarkMetaOption::class);
     $this->alterInfo('lark_entity_export_form_info');
     $this->setCacheBackend($cache_backend, 'lark_entity_export_form_plugins');
   }
@@ -27,7 +27,7 @@ final class EntityExportFormPluginManager extends DefaultPluginManager {
   /**
    * Get all plugin instances.
    *
-   * @return EntityExportFormInterface[]
+   * @return MetaOptionInterface[]
    */
   public function getInstances(array $configuration = []): array {
     $instances = [];
